+++ jest.config.js
@@ -53,7 +53,7 @@ module.exports = {
     ...packages.reduce(
       (acc, name) => ({
         ...acc,
-        '@bkui-vue/popover2': path.resolve(__dirname, './packages/popover2/src/index.ts'),
+        '@bkui-vue/popover': path.resolve(__dirname, './packages/popover/src/index.ts'),
         [`@bkui-vue/${name}(.*)$`]: path.resolve(__dirname, `./packages/${name}/src/$1`),
       }),
       {},
