+++ packages/bkui-vue/components.ts
@@ -23,6 +23,9 @@
 * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 * IN THE SOFTWARE.
 */
+import Popover2 from '@bkui-vue/popover';
+Popover2.name = 'Popover2';
+export { Popover2 };
 export { default as Alert }  from '@bkui-vue/alert';
 export { default as AnimateNumber  } from '@bkui-vue/animate-number';
 export { default as Affix  } from '@bkui-vue/affix';
@@ -41,7 +44,6 @@ export { default as Link  } from '@bkui-vue/link';
 export { default as Loading  } from '@bkui-vue/loading';
 export { default as Modal  } from '@bkui-vue/modal';
 export { default as Popover  } from '@bkui-vue/popover';
-export { default as Popover2 } from '@bkui-vue/popover2';
 export { default as Progress  } from '@bkui-vue/progress';
 export { default as Radio  } from '@bkui-vue/radio';
 export { default as Rate  } from '@bkui-vue/rate';
