+++ packages/dropdown/src/dropdown.less
@@ -17,7 +17,7 @@
     }
   }
 }
-.@{bk-prefix}-popover2 {
+.@{bk-prefix}-popover {
   &.@{bk-prefix}-pop2-content {
     &.@{bk-prefix}-dropdown-popover {
       padding: 0;
