+++ packages/popover/jest.config.js
@@ -28,5 +28,5 @@ const baseJestConf = require('../../jest.config');
 
 module.exports = {
   ...baseJestConf,
-  testRegex: 'packages/popover/__test__/.*\\.test\\.(js|ts|tsx)$',
+  testRegex: 'packages/popover-v2/__test__/.*\\.test\\.(js|ts|tsx)$',
 };
