+++ packages/popover/src/popover.less
@@ -1,7 +1,7 @@
 @import '@bkui-vue/styles/themes/themes.less';
 
 
-.@{bk-prefix}-popover2 {
+.@{bk-prefix}-popover {
   &.@{bk-prefix}-pop2-content {
     position: absolute;
     display: none;
