--- packages/popover2/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@bkui-vue/popover2",
-  "version": "0.0.0",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "clean": "rimraf dist && rimraf tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "yarn clean && tsc --build && cp \"./package.json\" ./dist/",
-    "test": "jest --passWithNoTests",
-    "lint": "eslint \"./src/**/*.{ts,tsx}\" --max-warnings=0"
-  },
-  "devDependencies": {
-    "@vue/test-utils": "2.1.0"
-  },
-  "peerDependencies": {
-    "vue": "^3.2.0"
-  }
-}
