+++ packages/select/src/select.less
@@ -254,7 +254,7 @@
   }
 }
 
-.@{bk-prefix}-popover2 {
+.@{bk-prefix}-popover {
   &.@{bk-prefix}-pop2-content {
     &.bk-select-popover {
       padding: 0;
