+++ packages/table/src/plugins/head-filter.less
@@ -17,7 +17,7 @@
   }
 }
 
-.@{bk-prefix}-popover2 {
+.@{bk-prefix}-popover {
   &.@{bk-prefix}-pop2-content {
     &.@{bk-prefix}-table-head-filter {
       padding: 5px 0 0 0;
@@ -27,6 +27,7 @@
       box-shadow: 0 3px 6px 0 rgba(0,0,0,.10);
 
       .content-list {
+        display: inline-block;
         max-width: 200px;
         min-width: 140px;
         overflow: auto;
