+++ packages/table/src/plugins/settings.less
@@ -4,6 +4,28 @@
   &[data-bk-table-settings-theme='true'] {
     padding: 0;
 
+    .bk-pop2-arrow {
+      &[data-arrow='top'] {
+        border-right: solid 1px #dcdee5;
+        border-bottom: solid 1px #dcdee5;
+      }
+
+      &[data-arrow='bottom'] {
+        border-top: solid 1px #dcdee5;
+        border-left: solid 1px #dcdee5;
+      }
+
+      &[data-arrow='left'] {
+        border-top: solid 1px #dcdee5;
+        border-right: solid 1px #dcdee5;
+      }
+
+      &[data-arrow='right'] {
+        border-bottom: solid 1px #dcdee5;
+        border-left: solid 1px #dcdee5;
+      }
+    }
+
     .setting-content {
       width: 515px;
       border: 1px solid #dcdee5;
