+++ packages/table/src/table.less
@@ -5,7 +5,7 @@
 @import '@bkui-vue/button/button.less';
 @import '@bkui-vue/exception/exception.less';
 @import '@bkui-vue/pagination/pagination.less';
-@import '@bkui-vue/popover2/popover2.less';
+@import '@bkui-vue/popover/popover.less';
 @import './plugins/head-filter.less';
 @import './plugins/head-sort.less';
 @import './plugins/settings.less';
